cancel Events.onKeyDown decoding w/Decode.fail
Use Decode.fail to short-circuit the Arrow key decoding on the Events.onKeyDown
subscription. This avoids having to explicitly handle the case where a key event
occurs that is not a direction arrow and generate a NoOp Msg. Using Decode.fail
no message is generated at all.

Haven't figured out any way to use same technique in the incoming port
swipeDirectionArrow which is processed by handleSwipe in the associated
subscription.

The swipeDirectionArrow is now a payload almost identical to that
produced by Events.onKeyDown:

  { key: <arrow-key-string> }