4 Move msg become one msg: Move (Maybe Direction)
The keyboard and swipe events received through subscriptions are
turned into Maybe Direction values which are now passed as an argument
to a single Move msg.

This also allows removal of the NoOp msg.

The data arriving at the incoming port swipeDirectionArrow
is now a payload almost identical to that produced
by Events.onKeyDown:

  { key: <arrow-key-string> }